Add files via upload
This was my first data science with python project. It was a guided project which investigated the netflix movies and guest stars in the office. This was after I completed the introduction/ starting the introduction to intermediate to Python from DataCamp. This project mostly focused on visualization of the data and extraction of data and filtering.